pyproject.toml
[tool.poetry]
name = "fiap-tech-challenge"
version = "0.1.0"
description = ""
authors = ["FelixMarcelo <felixmarcelo.mf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-decouple = "^3.8"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
requests = "^2.32.1"
beautifulsoup4 = "^4.12.3"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
